HHH-17770: Avoid casting long to int
This causes `NegativeArraySizeException: -1294967296` when Blob contents > 2Gb

+++ hibernate-core/src/main/java/org/hibernate/engine/jdbc/BlobProxy.java
@@ -201,7 +201,7 @@ public final class BlobProxy implements Blob, BlobImplementer {
 
 		@Override
 		public long getLength() {
-			return (int) length;
+			return length;
 		}
 
 		@Override
